requirements.txt
maafw==5.3.3
loguru
Pillow